+++ include/asbind20/ranges/tokenize_view.hpp
@@ -72,9 +72,11 @@ public:
             return *this;
         }
 
-        void operator++(int)
+        iterator operator++(int)
         {
+            iterator tmp = *this;
             next();
+            return tmp;
         }
 
         const value_type& operator*() const noexcept
@@ -115,6 +117,7 @@ public:
         }
     };
 
+    [[nodiscard]]
     bool empty() const noexcept
     {
         if(m_engine == nullptr) [[unlikely]]